Minor refactor and get balance
